+++ src/Ext/Rules/Body.h
@@ -10,12 +10,12 @@
 #include <Utilities/TemplateDef.h>
 #include <Utilities/Debug.h>
 
-
 class AnimTypeClass;
 class MouseCursor;
 class TechnoTypeClass;
 class VocClass;
 class WarheadTypeClass;
+class DigitalDisplayTypeClass;
 
 class RulesExt
 {
@@ -58,6 +58,10 @@ public:
 		Valueable<Point2D> Pips_SelfHeal_Infantry_Offset;
 		Valueable<Point2D> Pips_SelfHeal_Units_Offset;
 		Valueable<Point2D> Pips_SelfHeal_Buildings_Offset;
+		Valueable<Point2D> Pips_Generic_Size;
+		Valueable<Point2D> Pips_Generic_Buildings_Size;
+		Valueable<Point2D> Pips_Ammo_Size;
+		Valueable<Point2D> Pips_Ammo_Buildings_Size;
 
 		Valueable<bool> AllowParallelAIQueues;
 		Valueable<bool> ForbidParallelAIQueues_Aircraft;
@@ -82,6 +86,11 @@ public:
 		Valueable<bool> CrateOnlyOnLand;
 		Valueable<AffectedHouse> RadialIndicatorVisibility;
 		Valueable<bool> DrawTurretShadow;
+
+		ValueableVector<DigitalDisplayTypeClass*> Buildings_DefaultDigitalDisplayTypes;
+		ValueableVector<DigitalDisplayTypeClass*> Infantry_DefaultDigitalDisplayTypes;
+		ValueableVector<DigitalDisplayTypeClass*> Vehicles_DefaultDigitalDisplayTypes;
+		ValueableVector<DigitalDisplayTypeClass*> Aircraft_DefaultDigitalDisplayTypes;
 
 		Valueable<bool> AttachmentTopLayerMinHeight;
 		Valueable<bool> AttachmentUndergroundLayerMaxHeight;
@@ -114,6 +123,10 @@ public:
 			, Pips_SelfHeal_Infantry_Offset { { 25, -35 } }
 			, Pips_SelfHeal_Units_Offset { { 33, -32 } }
 			, Pips_SelfHeal_Buildings_Offset { { 15, 10 } }
+			, Pips_Generic_Size { { 4, 0 } }
+			, Pips_Generic_Buildings_Size { { 4, 2 } }
+			, Pips_Ammo_Size { { 4, 0 } }
+			, Pips_Ammo_Buildings_Size { { 4, 2 } }
 			, AllowParallelAIQueues { true }
 			, ForbidParallelAIQueues_Aircraft { false }
 			, ForbidParallelAIQueues_Building { false }
@@ -133,6 +146,10 @@ public:
 			, CrateOnlyOnLand { false }
 			, RadialIndicatorVisibility { AffectedHouse::Allies }
 			, DrawTurretShadow { false }
+			, Buildings_DefaultDigitalDisplayTypes {}
+			, Infantry_DefaultDigitalDisplayTypes {}
+			, Vehicles_DefaultDigitalDisplayTypes {}
+			, Aircraft_DefaultDigitalDisplayTypes {}
 			, AttachmentTopLayerMinHeight { 500 }
 			, AttachmentUndergroundLayerMaxHeight { -256 }
 		{ }