+++ src/Ext/Techno/Hooks.cpp
@@ -57,7 +57,112 @@ DEFINE_HOOK(0x6FA793, TechnoClass_AI_SelfHealGain, 0x5)
 	return SkipGameSelfHeal;
 }
 
-DEFINE_HOOK(0x70A4FB, TechnoClass_Draw_Pips_SelfHealGain, 0x5)
+DEFINE_HOOK(0x709B2E, TechnoClass_DrawPips_Sizes, 0x5)
+{
+	GET(TechnoClass*, pThis, ECX);
+	REF_STACK(int, pipWidth, STACK_OFFSET(0x74, -0x1C));
+
+	Point2D size;
+	bool isBuilding = pThis->WhatAmI() == AbstractType::Building;
+
+	if (pThis->GetTechnoType()->PipScale == PipScale::Ammo)
+	{
+		if (isBuilding)
+			size = RulesExt::Global()->Pips_Ammo_Buildings_Size;
+		else
+			size = RulesExt::Global()->Pips_Ammo_Size;
+
+		size = TechnoTypeExt::ExtMap.Find(pThis->GetTechnoType())->AmmoPipSize.Get(size);
+	}
+	else
+	{
+		if (isBuilding)
+			size = RulesExt::Global()->Pips_Generic_Buildings_Size;
+		else
+			size = RulesExt::Global()->Pips_Generic_Size;
+	}
+
+	pipWidth = size.X;
+	R->ESI(size.Y);
+
+	return 0;
+}
+
+DEFINE_HOOK(0x70A36E, TechnoClass_DrawPips_Ammo, 0x6)
+{
+	enum { SkipGameDrawing = 0x70A4EC };
+
+	GET(TechnoClass*, pThis, ECX);
+	LEA_STACK(RectangleStruct*, offset, STACK_OFFSET(0x74, -0x24));
+	GET_STACK(RectangleStruct*, rect, STACK_OFFSET(0x74, 0xC));
+	GET(int, pipWrap, EBX);
+	GET_STACK(int, pipCount, STACK_OFFSET(0x74, -0x54));
+	GET_STACK(int, maxPips, STACK_OFFSET(0x74, -0x60));
+	GET(int, yOffset, ESI);
+
+	auto const pTypeExt = TechnoTypeExt::ExtMap.Find(pThis->GetTechnoType());
+	Point2D position = { offset->X, offset->Y };
+
+	if (pipWrap > 0)
+	{
+		int levels = maxPips / pipWrap - 1;
+
+		for (int i = 0; i < pipWrap; i++)
+		{
+			int frame = pTypeExt->PipWrapAmmoPip;
+
+			if (levels >= 0)
+			{
+				int counter = i + pipWrap * levels;
+				int frameCounter = levels;
+				bool calculateFrame = true;
+
+				while (counter >= pThis->Ammo)
+				{
+					frameCounter--;
+					counter -= pipWrap;
+
+					if (frameCounter < 0)
+					{
+						calculateFrame = false;
+						break;
+					}
+				}
+
+				if (calculateFrame)
+					frame = frameCounter + frame + 1;
+			}
+
+			position.X += offset->Width;
+			position.Y += i * yOffset;
+
+			DSurface::Temp->DrawSHP(FileSystem::PALETTE_PAL, FileSystem::PIPS2_SHP,
+				frame, &position, rect, BlitterFlags(0x600), 0, 0, ZGradient::Ground, 1000, 0, 0, 0, 0, 0);
+		}
+	}
+	else
+	{
+		int ammoFrame = pTypeExt->AmmoPip;
+		int emptyFrame = pTypeExt->EmptyAmmoPip;
+
+		for (int i = 0; i < maxPips; i++)
+		{
+			if (i >= pipCount && emptyFrame < 0)
+				break;
+
+			int frame = i >= pipCount ? emptyFrame : ammoFrame;
+			position.X += offset->Width;
+			position.Y += i * yOffset;
+
+			DSurface::Temp->DrawSHP(FileSystem::PALETTE_PAL, FileSystem::PIPS2_SHP,
+				frame, &position, rect, BlitterFlags(0x600), 0, 0, ZGradient::Ground, 1000, 0, 0, 0, 0, 0);
+		}
+	}
+
+	return SkipGameDrawing;
+}
+
+DEFINE_HOOK(0x70A4FB, TechnoClass_DrawPips_SelfHealGain, 0x5)
 {
 	enum { SkipGameDrawing = 0x70A6C0 };